% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMissingElements.R
\name{addMissingElements}
\alias{addMissingElements}
\title{Add Missing Elements}
\usage{
addMissingElements(data, standParams)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{standParams}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}
}
\value{
The same data.table as what was passed ("data") but possibly with
  additional rows.
}
\description{
This function takes a data.table and adds rows to it so that the data 
contains an observation for all elements for each commodity in the dataset.
}
