% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFoodAggregates.R
\name{calculateFoodAggregates}
\alias{calculateFoodAggregates}
\title{calculateFoodAggregates}
\usage{
calculateFoodAggregates(standData = standData, params = c(),
  yearVals = yearVals)
}
\arguments{
\item{standData}{The data.frame resulting from the standardization process
containing all the elements needed for the calculation of DES}

\item{params}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}
}
\value{
A data.table containing all the food aggregates, by country-commodity-year
and also containing the population by year
}
\description{
The function creates all the Food aggregates to be displayed in the FBS.
This function calls the popoulation dataset in the SWS and displays population
in the resulting datatable
}

