% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkShareValue.R
\name{checkShareValue}
\alias{checkShareValue}
\title{Check if values of shares are consistent and substitute inconsistent values}
\usage{
checkShareValue(tree2Subset)
}
\arguments{
\item{One}{single subset of the Tree, which means, a subset having all the parent of a single child
where the child is an official one. The subset has the characteristis of having shares not summing at 1}
}
\value{
The function returns the subTree with the corrected shares and
with a severity index to be saved back in order to have a different color also where there 
are shares that have been changed
}
\description{
This function checks between a subset of the Tree having official shares 
but with the sum not equal to 1. 
It checks where is the problem and correct it
The structure of this function is articulated on purpose
for sake of clarity
}
