% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseEdges.R
\name{collapseEdges}
\alias{collapseEdges}
\title{Collapse Edges}
\usage{
collapseEdges(edges, parentName = "parentID", childName = "childID",
  extractionName = "extractionRate", keyCols = c("timePointYearsSP",
  "geographicAreaFS"))
}
\arguments{
\item{edges}{A data.table containing three columns (corresponding to a 
parent, a child, and an extraction rate).}

\item{parentName}{The column name of the parent column in edges.}

\item{childName}{The column name of the child column in edges.}

\item{extractionName}{The column name of the extraction rate column in 
edges.}

\item{keyCols}{The column name(s) of the columns of edges which should be
considered as keys.  These columns will thus be included in any joins to
ensure we don't get duplicates.  If there is no key, this can be "".}
}
\value{
A data.table of the same structure as edges, but with intermediate 
  steps removed.
}
\description{
This function takes a "edge" data.table (i.e. a data.table with parent/child
columns and an extraction rate column) and condenses the parent/child 
relationships.  For example, wheat may be processed into flour which is in 
turn processed into bread.  Thus, we can standardize bread to flour and then
the wheat, but it'd be easier to just standardize straight to bread.  This 
function condenses the edges by multiplying the extraction rates so that you
can take bread straight to wheat.
}

