% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeFbsAggregate.R
\name{computeFbsAggregate}
\alias{computeFbsAggregate}
\title{Compute FBS Aggregate}
\usage{
computeFbsAggregate(data, fbsTree, standParams)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{fbsTree}{This "tree" should just have three columns:
standParams$parentID, standParams$childID, and standParams$extractVar
(which if missing will just be assigned all values of 1).  This tree
specifies how SUA commodities get combined to form the FBS aggregates.  If
NULL, the last step (aggregation to FBS codes) is skipped and data is
simply returned at SUA level.}

\item{standParams}{The parameters for standardization.  These parameters
  provide information about the columns of data and tree, specifying (for
  example) which columns should be standardized, which columns represent
  parents/children, etc.}
}
\value{
A list of four data.tables, each containing the final FBS data at
  the four different FBS levels.
}
\description{
This function takes the FBS aggregation tree and rolls up the SUA data to
all FBS levels.
}

