% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSupplyComponents.R
\name{computeSupplyComponents}
\alias{computeSupplyComponents}
\title{Final Standardization to Primary Equivalent}
\usage{
computeSupplyComponents(data, standParams, loop)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{standParams}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}

\item{loop}{just an index of loops}
}
\value{
A data.table with the aggregated primary commodities.
}
\description{
After the full SUA has been balanced, all the commodities need to be rolled 
up to their primary equivalents.  This function does this, aggregating up 
trade and food.
}

