% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutItems}
\alias{cutItems}
\title{co-products of the processing chain which do NOT have to be back Standardized for
not to create double counting}
\format{a vector of 90 cpc codes}
\usage{
cutItems
}
\description{
co-products of the processing chain which do NOT have to be back Standardized for
not to create double counting
}
\keyword{datasets}
