% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDownloadFix.R
\name{dataDownloadFix}
\alias{dataDownloadFix}
\title{Data Download and Fix of sugar Values}
\usage{
dataDownloadFix(key = c(), p = c())
}
\arguments{
\item{key}{Key with dimensions for data download.}

\item{p}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}
}
\value{
A data.table with the data fixed for sugar
}
\description{
This function download the data from sua_unbalanced dataset, then harmonize sugar codes. 
Raw cane and beet sugar are considered as separate codes in some domain, like trade, because sugar raw can be traded
as beet raw, cane raw or just raw (23511.01, 23512 or 2351f), 
but when one has to go from the processed product to the primary product, 
is not possible to know if a code 2351f has to be standardized to cane or beet, therefore 
in the standardization process cane and beet have to be considered as a unique code (2351f)
This function makes checks and harmonize sugar codes
}
