% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementCodesToNames.R
\name{elementCodesToNames}
\alias{elementCodesToNames}
\title{Element Codes to Names}
\usage{
elementCodesToNames(data, elementCol = NULL, itemCol = NULL,
  standParams = NULL)
}
\arguments{
\item{data}{The data.table containing the codes to be mapped.  It is assumed
to have one column with element codes and one column with item codes.}

\item{elementCol}{The column name of the column with element codes.}

\item{itemCol}{The column name of the column with item codes.}

\item{standParams}{A list of parameters (see 
faoswsStandardization::defaultStandardizationParameters).  If supplied, 
elementCol and itemCol can be NULL.}
}
\value{
The passed data.table "data" but with an updated element column
  (with names instead of codes).
}
\description{
This function converts element codes (such as 5510, 5421, ...) into 
meaningful names (such as "production", "yield", etc.).  It only works with 
the codes defined in a file called elementCodes, which was om the Share drive
paste0(R_SWS_SHARE_PATH,"/browningj/") but now is an ad hoc dataTable in the SWS
in the AgricultureProduction domain.  The tricky part of this function is that
the commodity type must be used to correctly choose the element code, as 
different commodities may use different codes for production, yield, ...
}

