% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbsTree.R
\docType{data}
\name{fbsTree}
\alias{fbsTree}
\title{FBS Tree}
\format{A data table with 264 rows and 5 variables}
\usage{
data(fbsTree)
}
\description{
This dataset contains the aggregation rules for commodities up into FBS
aggregates.  Ultimately, this table should probably exist on the SWS.
}
\details{
\itemize{
     \item measuredItemSuaFbs The commodity code, in CPC format.
     \item fbsID4 The lowest FBS aggregate code.  On the disseminated food
     balance sheets, data is provided at the level of this code, as well as
     aggregated levels of this code.  For example, see 
     \url{http://faostat3.fao.org/download/FB/FBS/E}.
     \item fbsID3 The third-level FBS code.
     \item fbsID2 The second-level FBS code.
     \item fbsID1 The top-level FBS code.
}
}
\keyword{datasets}

