% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalStandardizationToPrimary.R
\name{finalStandardizationToPrimary}
\alias{finalStandardizationToPrimary}
\title{Final Standardization to Primary Equivalent}
\usage{
finalStandardizationToPrimary(data, tree, standParams, sugarHack = TRUE,
  specificTree = TRUE, cut = c(), additiveElements = c())
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{tree}{The commodity tree which provides the edge structure.  Note that
this tree may be different than the processing tree, in particular if some
commodities will be standardized into a different aggregate.}

\item{standParams}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}

\item{sugarHack}{Logical.  See standardizeTree for details.}

\item{specificTree}{Logical.  Is a country/year specific commodity tree 
being provided?}

\item{additiveElements}{Column names of data which should be
aggregated/standardized via simple addition.}
}
\value{
A data.table with the aggregated primary commodities.
}
\description{
After the full SUA has been balanced, all the commodities need to be rolled 
up to their primary equivalents.  This function does this, aggregating up 
trade and food.
}
