% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFBSTree.R
\name{getFBSTree}
\alias{getFBSTree}
\title{Get FBS Tree}
\usage{
getFBSTree()
}
\value{
A data.table object with the FBS tree.
}
\description{
This function creates the tree for aggregation to FBS commodities.
}

