% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOldCommodityTree.R
\name{getOldCommodityTree}
\alias{getOldCommodityTree}
\title{Get Old Commodity Tree}
\usage{
getOldCommodityTree(geographicAreaFS, timePointYears)
}
\arguments{
\item{geographicAreaFS}{A character vector containing the geographic area
codes in FS format.}

\item{timePointYears}{A character vector containing the years for which
commodity trees are wanted.}
}
\value{
A data.table of the commodity tree.
}
\description{
This function pulls the old commodity trees from the SWS.  These trees are
constructed from the extraction rate and shares data from the AUPUS domain.
}
