% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustCommodityTree.R
\name{huber}
\alias{huber}
\title{Wrapper for function from MASS}
\usage{
huber(y, ...)
}
\arguments{
\item{y}{The vector of data.}

\item{...}{Additional arguments to pass to MASS::huber}
}
\value{
A numeric vector of length two with the estimated mean and standard
  deviation.
}
\description{
This function is designed to call MASS::huber with two exceptions: missing 
values are removed and, if MASS::huber results in an error, the non-robust 
mean/sd are returned.
}

