% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualTreeCorrections.R
\name{manualTreeCorrections}
\alias{manualTreeCorrections}
\title{This function makes some manual corrections on TREE
for all those cases in which there is not an extraction rate}
\usage{
manualTreeCorrections(tree = tree)
}
\arguments{
\item{tree}{the Tree to be modified}

\item{cereals}{if to modify or not cereals. 
this has been added because some changes on cereals were tested as bad}
}
\value{
the tree updated
}
\description{
This function makes some manual corrections on TREE
for all those cases in which there is not an extraction rate
}
