% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/markUpdated.R
\name{markUpdated}
\alias{markUpdated}
\title{Mark Updated Cells}
\usage{
markUpdated(new, old, standParams)
}
\arguments{
\item{new}{The new data.table with the standardization data.}

\item{old}{The original data.table with the standardization data.}

\item{standParams}{The parameters for standardization.  These parameters
  provide information about the columns of data and tree, specifying (for
  example) which columns should be standardized, which columns represent
  parents/children, etc.}
}
\value{
The same data.table as new but with an additional column.
}
\description{
This function compares two data.tables for differences.  Note that it
assumes a very specific structure, as it is intended to be used with the
standardization code only.  An additional column is placed on the first
data.table that is named updateFlag, and it is TRUE if there is a difference
and FALSE otherwise.
}

