% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCommodityTrees.R
\name{plotCommodityTrees}
\alias{plotCommodityTrees}
\title{Plot Commodity Trees}
\usage{
plotCommodityTrees(commodityTree, parentColname, childColname,
  extractionColname, dir = getwd(), prefix = "commodity_tree",
  adjExtractionRates = TRUE)
}
\arguments{
\item{commodityTree}{A data.table with parent and child node IDs
(corresponding to the IDs in nodes) which specify the commodity tree
structure.  Additionally, there should be a column with extraction rate
data and a column with shares data.}

\item{parentColname}{The column name of commodityTree which contains the ID
of the parent node.}

\item{childColname}{The column name of commodityTree which contains the ID
of the child node.}

\item{extractionColname}{The column name of commodityTree which contains
the extraction rate data.}

\item{dir}{The directory where the commodity trees should be saved.}

\item{prefix}{The prefix to add to the name of the plot(s) which will be
saved to dir.  One plot is created for each commodity tree.}

\item{adjExtractionRates}{Logical.  Typically, the extraction rates are
stored as percentages multiplied by 10000, i.e. 5% is represented as 
0.05*10000 = 500.  If TRUE, the function converts them to percentages.}
}
\value{
No object is returned, but image files of the plots are written
out to .png files.  The input parameter dir specifies where these files
will be placed.
}
\description{
This function generates plots for each of the commodity trees defined
in a commodity tree object.  A separate plot is given for each node that
has no parent, and this should allow for the production of a wheat tree,
a milk tree, etc.
}
