% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingleTree.R
\name{plotSingleTree}
\alias{plotSingleTree}
\title{Plot Single Tree}
\usage{
plotSingleTree(edges, parentColname, childColname, extractionColname,
  dashInf = TRUE, ...)
}
\arguments{
\item{edges}{A data.table with parent and child node IDs (corresponding to 
the IDs in nodes) which specify the commodity tree structure. 
Additionally, there should be a column with extraction rate data and a 
column with shares data.}

\item{parentColname}{The column name of commodityTree which contains the ID 
of the parent node.}

\item{childColname}{The column name of commodityTree which contains the ID 
of the child node.}

\item{extractionColname}{The column name of commodityTree which contains the
extraction rate data.}

\item{dashInf}{Should extraction rates of infinity be represented with
dashes?  This generally makes sense because such processing should be
shown on the tree but not considered in standardization.}

\item{...}{Additional plotting parameers for diagram::plotmat.}
}
\value{
Generates a plot of the commodity tree as specified by edges.
}
\description{
This function generates a plot for one commodity trees defined with an edge 
list.
}
