% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSUATableNames.R
\name{printSUATableNames}
\alias{printSUATableNames}
\title{Print SUA Table with name of commodities}
\usage{
printSUATableNames(data, standParams, printCodes, printProcessing = TRUE,
  nutrientElements = c())
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{standParams}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}

\item{printCodes}{A character vector of the elements of interest.  This is
required to keep the table from getting to big/unreadable.}

\item{printProcessing}{Logical.  Should food processing also be printed?}

\item{nutrientElements}{A list of the nutrient codes which should also be
printed.}
}
\value{
Nothing is returned, but a table is returned in a nice format to 
  show the current SUA data.
}
\description{
Print SUA Table with name of commodities
}

