% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processForward.R
\name{processForward}
\alias{processForward}
\title{Process Forward}
\usage{
processForward(data, tree, standParams)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{tree}{The commodity tree which provides the edge structure.}

\item{standParams}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}
}
\value{
A list of names 'data' and 'tree'.  Both objects must be returned, 
  as the tree is updated by pruning off some edges.
}
\description{
A few edges in the commodity trees are labeled with an 'F' to indicate that 
processing is 'forward'.  The parent commodities in these edges are 
immediately converted into the corresponding child and then they are removed
from the tree (as we will standardize to the children instead).  This is a 
rare scenario; an example commodity is sugar.
}
\details{
Note: when commodities are processed forward like this, the final flag is
assigned ARBITRARILY as the first flag observed (for lack of a better
approach).  This should be corrected.
}

