% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSUATable.R, R/printSUATableNames.R
\name{roundNum}
\alias{roundNum}
\title{Round numbers}
\usage{
roundNum(x)

roundNum(x)
}
\arguments{
\item{x}{A number to be nicely formatted.}

\item{x}{A number to be nicely formatted.}
}
\value{
The number as a character string, formatted "nicely".

The number as a character string, formatted "nicely".
}
\description{
Helper function for rounding numbers for display.

Helper function for rounding numbers for display.
}

