% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFBSItermediateStep.R
\name{saveFBSItermediateStep}
\alias{saveFBSItermediateStep}
\title{Save Intermediate steps of Standardization process}
\usage{
saveFBSItermediateStep(directory, fileName, data)
}
\arguments{
\item{directory}{Where does the intermediate step have to be saved back to?}

\item{fileName}{Name of the file we are creating}

\item{data}{data.table to be saved}
}
\description{
This function has been created in order to locally save some intermediate
steps of the Standardization module. This exigency arose for validation
purposes.
}
