% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendMail4Forced.R
\name{sendMail4forced}
\alias{sendMail4forced}
\title{Send email after having checked if some protected figure of production has been changed}
\usage{
sendMail4forced(forcedData)
}
\arguments{
\item{forcedData}{If existing, are the SUA of the commodity for which
the change of the production figure was required in order to Fill the SUA}
}
\value{
The function doesn't return anything, but send mail
}
\description{
This function simply snd email to the user, 
with the result of the check on shares
}

