% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendMail4shares.R
\name{sendMail4shares}
\alias{sendMail4shares}
\title{Send email after having checked for shares}
\usage{
sendMail4shares(tree)
}
\arguments{
\item{tree}{single subset of the Tree, which means, a subset having all the parent of a single child
where the child is an official one. The subset has the characteristis of having shares not summing at 1}
}
\value{
The function doesn't return anything, but send mail
}
\description{
This function simply snd email to the user, 
with the result of the check on shares
}
