% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendMailAttachment.R
\name{sendMailAttachment}
\alias{sendMailAttachment}
\title{Send email after having checked if some protected figure of production has been changed}
\usage{
sendMailAttachment(fileToSend, name, textBody)
}
\arguments{
\item{fileToSend}{If existing, are the SUA of the commodity for which
the change of the production figure was required in order to Fill the SUA}
}
\value{
The function doesn't return anything, but send mail
}
\description{
This function simply snd email to the user, 
with the result of the check on shares
}

