% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizationWrapper_validation.R
\name{standardizationWrapper_validation}
\alias{standardizationWrapper_validation}
\title{FULL STANDARDIZATION PROCESS - VERSION FOR VALIDATION OF OLD DATA}
\usage{
standardizationWrapper_validation(data, tree, fbsTree = NULL, standParams,
  nutrientData = NULL, printCodes = c(), debugFile = NULL,
  batchnumber = batchnumber, utilizationTable = utilizationTable)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization. 
It should have columns corresponding to year, country, element, commodity,
and value.  The specific names of these columns should be in standParams.}

\item{tree}{The commodity tree which details how elements can be processed 
into other elements.  It does not, however, specify which elements get 
aggregated into others.  This data.table should have columns parent, 
child, extraction rate, share, and target (target specifying if an element
is processed forward or not).  Names of these columns should be provided
in standParams.}

\item{fbsTree}{This "tree" should just have three columns: 
standParams$parentID, standParams$childID, and standParams$extractVar 
(which if missing will just be assigned all values of 1).  This tree 
specifies how SUA commodities get combined to form the FBS aggregates.  If
NULL, the last step (aggregation to FBS codes) is skipped and data is 
simply returned at SUA level.}

\item{standParams}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}

\item{nutrientData}{A data.table containing one column with the item codes 
(and this column's name must match standParams$itemVar) and additional 
columns representing nutrient information.  For example, you could have 4 
columns: measuredItemCPC, calories, proteins, fats.  In the calories, 
proteins, and fats columns there should be numeric values representing 
multipliers to convert kilograms of the item into calories/proteins/fats. 
If NULL, nothing is done with nutrients.}

\item{printCodes}{A list of the item codes which should be printed at each 
step to show the progress of the algorithm.}

\item{debugFile}{folder for saving the intermediate files.}

\item{batchnumber}{Number of batch running.}

\item{utilizationTable}{Table of utilization for suaFilling}
}
\value{
A data.table containing the final balanced and standardized SUA 
  data.  Additionally, this table will have new elements in it if 
  nutrientData was provided.
}
\description{
This version is designed for working on Sua data coming from old methodology
}
\details{
This process differs from the "regular" one for: 
- utilization table. In this case loss rank is kept because in the old mthodology there was loss
  also for non primary products. Loss is then been deleted in the new methodology version
  
- Sua Filling is slightly different
}
