% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suaFilling_validation.R
\name{suaFilling_validation}
\alias{suaFilling_validation}
\title{FUNCTION FOR FILLING THE SUA FOR MISSING OR WRONG VALUES.
THIS IS THE VERSION USED FOR VALIDATION OF DATA FROM OLD METHODOLOGY}
\usage{
suaFilling_validation(data, p = p, tree = tree, primaryCommodities = c(),
  stockCommodities = c(), debugFile = NULL, utilizationTable = c(),
  imbalanceThreshold = 10, loop1 = TRUE)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{p}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}

\item{tree}{this is the sub tree used in the function.}

\item{primaryCommodities}{Primary level commodities (such as wheat, oranges,
sweet potatoes, etc.) should not be balanced at this step but rather by
the balancing algorithm.  This argument allows the user to specify a
character vector with these primary element codes.}

\item{stockCommodities}{This list specify if the commodity can be allocated to stock}

\item{utilizationTable}{is the external utilizataion table}

\item{imbalanceThreshold}{The size that the imbalance must be in order for 
an adjustment to be made.}

\item{loop1}{insicates if is the firs or second loop before or after the food Proc calculation}

\item{cut}{these are primary equivalent commodities.}
}
\value{
the Value column of the passed data.table is updated
}
\description{
This function replaces the old BalanceResidual
}
\details{
This function forces a the filling of empty elements in the pulled SUA
by allocating the "imbalance" according to a Ranking of the possible 
Uitlizazions for each combination of country/commodity

- If supply < utilization
  * the imbalance is assigned to production, if it is not official
  * if is official it is checked if the commodity is a stock commodity
    i.e. if it does make sense to allocate stock. If it is so, stock is created (negative)
  * if production is official and is not a stocj comodity, 
    Utilizations are proportionally reduced,if the difference S-U <=10% Utilizations
  * if none of these operations is possible, a manual check of the data is needed
    (for the moment production is increased anyway and an external file is produced for check)
- if supply > utilization a Multiple filler approach is used: 
  * If all the ranked utilizations are present 
    these are proportionally incremented
  * if one is empty, imbalance goes to this commodity
  * if more than one is empty, Utilization are created based on a ranking of possible Utilization 
    coming from an external source.
    The approach used is of the Inverse Ranking.

Trade and Food Processing is never created in this function:
- trade is supposed to be "protected", 
- food processing is created outside this function 
  using a specific function and is not replaced here
}
