% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaWheat.R
\docType{data}
\name{usaWheat}
\alias{usaWheat}
\title{USA wheat SUA from 2011}
\format{A data table with 22 rows and 7 variables}
\usage{
data(usaWheat)
}
\description{
This dataset contains an example of the data that could occur in the SUA
file.  Note that, at the time of creation, this data represents all the data
that was currently available for USA wheat in the SUA table, but that more
data would become available as other modules are run.  Thus, this dataset is
intended to be used ONLY as an example of how the code works.
}
\details{
\itemize{
     \item geographicAreaM49 The geographic area code for USA.
     \item measuredElementSuaFbs The element code.  5510 represents
     production, for example.
     \item measuredItemSuaFbs The commodity code, in CPC format.  For
     example, 0111 corresponds to wheat.
     \item timePointYears The year.
     \item Value The value corresponding to the particular commodity/element
     for the USA in 2011.
     \item flagObservationStatus The first flag, indicating the status of the
     observation.
     \item flagMethod The second flag, providing more detail on the
     collection/imputation method for the observations.
}
}
\keyword{datasets}

