% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeroWeight}
\alias{zeroWeight}
\title{Table of possible utilization by country/commodity combinations}
\format{a data.table with 155043 rows and 5 columns
\itemize{
 \item geographicAreaM49, M49 code of the country
 \item measuredElementSuaFbs, utilization Element
 \item measuredItemSuaFbs, CPC code
 \item rank, rank of that element
 \item rankInv, inverse rank of that element
 }}
\usage{
zeroWeight
}
\description{
THIS TABLE EXCLUDES LOSS FOR DERIVED COMMODITIES
}
\details{
Utilization have a rank representing the importance of that element, 
in terms of mean value over all the years for that combination
}
\keyword{datasets}
