% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceResidual.R
\name{balanceResidual}
\alias{balanceResidual}
\title{Balance Residual}
\usage{
balanceResidual(data, standParams, feedCommodities = c(),
  indCommodities = c(), primaryCommodities = c(),
  foodProcessCommodities = c(), imbalanceThreshold = 10)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{standParams}{The parameters for standardization.  These parameters 
provide information about the columns of data and tree, specifying (for 
example) which columns should be standardized, which columns represent 
parents/children, etc.}

\item{feedCommodities}{Sometimes excess supply will need to be allocated to 
some processed product.  The default is to place it into food, but this 
list specifies which elements should allocate such a difference to feed.}

\item{indCommodities}{Same as feedCommodities, but for commodities where we 
allocate the difference to industrial utilization.}

\item{primaryCommodities}{Primary level commodities (such as wheat, oranges,
sweet potatoes, etc.) should not be balanced at this step but rather by
the balancing algorithm.  This argument allows the user to specify a
character vector with these primary element codes.}

\item{foodProcessCommodities}{Same as feedCommodities, but for commodities 
where we allocate the difference to food processing.}

\item{imbalanceThreshold}{The size that the imbalance must be in order for 
an adjustment to be made.}
}
\value{
Nothing is returned, but the Value column of the passed data.table 
  is updated.
}
\description{
This function forces a balance in the passed elements by allocating the 
imbalance to one element.
}
\details{
If supply < utilization, the imbalance is always assigned to production (as 
trade is generally assumed to be fixed and stock changes are usually 0 and 
hence also fixed).

If supply > utilization, we must choose where to allocate the imbalance. The
default variable is food, but for some commodities we could instead use 
feed, food processing, or industrial.
}

