% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultStandardizationParameters.R
\name{defaultStandardizationParameters}
\alias{defaultStandardizationParameters}
\title{Default Standardization Parameters}
\usage{
defaultStandardizationParameters()
}
\value{
A list with the standardization parameters.
}
\description{
Provides an object which contains the standardization parameters.  This
allows for easy passing into functions.
}

