% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoswsStandardization-package.R
\docType{package}
\name{faoswsStandardization-package}
\alias{faoswsStandardization}
\alias{faoswsStandardization-package}
\title{Package for standardization of Food Balance Sheets}
\description{
Package for standardization of Food Balance Sheets
}
\author{
Josh Browning <joshua.browning@fao.org>
}

